+++ test/OpenTelemetry.Exporter.Zipkin.Tests/ZipkinExporterTests.cs
@@ -554,7 +554,9 @@ public class ZipkinExporterTests : IDisposable
 
         if (status.HasValue)
         {
+#pragma warning disable
             activity.SetStatus(status.Value);
+#pragma warning enable
         }
 
         activity.SetEndTime(endTimestamp);